refactor: `FundingStatus`, 'FriendshipStatus' 모듈 변경
- `FundingStatus` : core -> share
- 'FriendshipStatus' : share -> member
