+++ bc/core/src/main/java/app/giftify/funding/adpater/inbound/web/FundingController.java
@@ -2,10 +2,10 @@ package app.giftify.funding.adpater.inbound.web;
 
 import app.giftify.funding.adpater.inbound.dto.*;
 import app.giftify.funding.application.FundingFacade;
-import app.giftify.funding.domain.FundingStatus;
 import app.giftify.security.common.CurrentMemberId;
 import app.giftify.shared.api.paging.PageResponse;
 import app.giftify.shared.api.response.RsData;
+import app.giftify.shared.domain.type.FundingStatus;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
