+++ bc/core/src/main/java/app/giftify/funding/adpater/inbound/web/FundingV2ApiSpec.java
@@ -1,10 +1,10 @@
 package app.giftify.funding.adpater.inbound.web;
 
 import app.giftify.funding.adpater.inbound.dto.*;
-import app.giftify.funding.domain.FundingStatus;
 import app.giftify.security.common.CurrentMemberId;
 import app.giftify.shared.api.paging.PageResponse;
 import app.giftify.shared.api.response.RsData;
+import app.giftify.shared.domain.type.FundingStatus;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
