+++ bc/core/src/main/java/app/giftify/funding/adpater/outbound/jpa/Funding.java
@@ -1,8 +1,8 @@
 package app.giftify.funding.adpater.outbound.jpa;
 
-import app.giftify.funding.domain.FundingStatus;
 import app.giftify.funding.domain.exception.FundingErrorCode;
 import app.giftify.funding.domain.exception.FundingException;
+import app.giftify.shared.domain.type.FundingStatus;
 import app.giftify.support.jpa.BaseJpaEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
