+++ bc/core/src/main/java/app/giftify/funding/adpater/outbound/repository/FundingParticipantMemberRepository.java
@@ -3,7 +3,7 @@ package app.giftify.funding.adpater.outbound.repository;
 import app.giftify.funding.adpater.outbound.jpa.Funding;
 import app.giftify.funding.adpater.outbound.jpa.FundingParticipantMember;
 import app.giftify.funding.application.MyFundingInfo;
-import app.giftify.funding.domain.FundingStatus;
+import app.giftify.shared.domain.type.FundingStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
