+++ bc/core/src/main/java/app/giftify/funding/application/FundingCloseUseCase.java
@@ -4,11 +4,11 @@ import app.giftify.funding.adpater.inbound.dto.FundingCompleteResponseDto;
 import app.giftify.funding.adpater.outbound.jpa.Funding;
 import app.giftify.funding.adpater.outbound.repository.FundingParticipantMemberRepository;
 import app.giftify.funding.adpater.outbound.repository.FundingRepository;
-import app.giftify.funding.domain.FundingStatus;
 import app.giftify.funding.domain.exception.FundingErrorCode;
 import app.giftify.funding.domain.exception.FundingException;
 import app.giftify.shared.domain.event.EventPublisher;
 import app.giftify.shared.domain.event.funding.FundingCanceledEvent;
+import app.giftify.shared.domain.type.FundingStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
