+++ bc/core/src/main/java/app/giftify/funding/application/GetFundingSnapshotUseCase.java
@@ -1,7 +1,7 @@
 package app.giftify.funding.application;
 
 import app.giftify.funding.adpater.outbound.repository.FundingRepository;
-import app.giftify.funding.domain.FundingStatus;
+import app.giftify.shared.domain.type.FundingStatus;
 import app.giftify.shared.domain.vo.FundingSnapshot;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
