+++ bc/core/src/main/java/app/giftify/funding/application/SyncFundingProductUseCase.java
@@ -2,7 +2,7 @@ package app.giftify.funding.application;
 
 import app.giftify.funding.adpater.outbound.jpa.Funding;
 import app.giftify.funding.adpater.outbound.repository.FundingRepository;
-import app.giftify.funding.domain.FundingStatus;
+import app.giftify.shared.domain.type.FundingStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
