+++ bc/member/src/main/java/app/giftify/friendship/application/service/FriendshipService.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import app.giftify.friendship.application.port.in.*;
 import app.giftify.friendship.application.port.out.FriendshipRepositoryPort;
 import app.giftify.friendship.domain.Friendship;
-import app.giftify.shared.domain.type.FriendshipStatus;
+import app.giftify.friendship.domain.FriendshipStatus;
 import app.giftify.friendship.domain.exception.FriendshipErrorCode;
 import app.giftify.friendship.domain.exception.FriendshipException;
 import app.giftify.member.application.port.out.MemberRepositoryPort;
